src/main/java/com/example/demo/service/impl/UserRoleServiceImpl.java
package com.example.demo.service.impl;

import com.example.demo.model.entity.UserRoleEntity;
import com.example.demo.repository.UserRoleRepository;
import com.example.demo.service.UserRoleService;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class UserRoleServiceImpl implements UserRoleService {
    private final UserRoleRepository userRoleRepository;

    public UserRoleServiceImpl(UserRoleRepository userRoleRepository) {
        this.userRoleRepository = userRoleRepository;
    }

    @Override
    public List<UserRoleEntity> getAllRoles() {
        return userRoleRepository.findAll();
    }
}
